+++ PublicKeyInfrastructure/src/main/java/com/pki/example/service/CSRService.java
@@ -4,8 +4,8 @@ import com.pki.example.data.CARequest;
 import com.pki.example.data.CertificateResponse;
 
 public interface CSRService {
-    public CertificateResponse approveCSR(Long csrId, Long issuerUserId, String issuerAlias) throws Exception;
+    public CertificateResponse approveCSR(Long csrId, Integer issuerUserId, String issuerAlias) throws Exception;
 
-    Long saveCSR(org.springframework.web.multipart.MultipartFile file, Long userId) throws Exception;
+    Long saveCSR(org.springframework.web.multipart.MultipartFile file, Integer userId) throws Exception;
 
 }
